test: Move common resharding functions into base_resharding.py.
Fixed two pylint errors in resharding.py.
