+++ test/initial_sharding_bytes.py
@@ -4,13 +4,16 @@
 # Use of this source code is governed by a BSD-style license that can
 # be found in the LICENSE file.
 
+"""Re-runs initial_sharding.py with a varbinary keyspace_id."""
+
 from vtdb import keyrange_constants
 
+import base_sharding
 import initial_sharding
 import utils
 
 # this test is just re-running an entire initial_sharding.py with a
 # varbinary keyspace_id
 if __name__ == '__main__':
-  initial_sharding.keyspace_id_type = keyrange_constants.KIT_BYTES
+  base_sharding.keyspace_id_type = keyrange_constants.KIT_BYTES
   utils.main(initial_sharding)
