+++ test/merge_sharding.py
@@ -13,8 +13,6 @@ Note this test is just testing the full workflow, not corner cases or error
 cases. These are mostly done by the other resharding tests.
 """
 
-import struct
-
 import logging
 import unittest
 
@@ -25,9 +23,6 @@ import environment
 import tablet
 import utils
 
-
-keyspace_id_type = keyrange_constants.KIT_UINT64
-pack_keyspace_id = struct.Struct('!Q').pack
 
 # initial shards
 # shard -40
@@ -84,7 +79,7 @@ class TestMergeSharding(unittest.TestCase, base_sharding.BaseShardingTest):
   # create_schema will create the same schema on the keyspace
   # then insert some values
   def _create_schema(self):
-    if keyspace_id_type == keyrange_constants.KIT_BYTES:
+    if base_sharding.keyspace_id_type == keyrange_constants.KIT_BYTES:
       t = 'varbinary(64)'
     else:
       t = 'bigint(20) unsigned'
@@ -112,75 +107,6 @@ index by_msg (msg)
                      '-sql=' + create_view_template % ('view1', 'resharding1'),
                      'test_keyspace'],
                     auto_log=True)
-
-  # _insert_value inserts a value in the MySQL database along with the comments
-  # required for routing.
-  def _insert_value(self, tablet_obj, table, mid, msg, keyspace_id):
-    k = utils.uint64_to_hex(keyspace_id)
-    tablet_obj.mquery(
-        'vt_test_keyspace',
-        ['begin',
-         'insert into %s(id, msg, keyspace_id) '
-         'values(%d, "%s", 0x%x) /* vtgate:: keyspace_id:%s */ '
-         '/* id:%d */' %
-         (table, mid, msg, keyspace_id, k, mid),
-         'commit'],
-        write=True)
-
-  def _get_value(self, tablet_obj, table, mid):
-    """Returns the row(s) from the table for the provided id, using MySQL.
-
-    Args:
-      tablet_obj: the tablet to get data from.
-      table: the table to query.
-      mid: id field of the table.
-    Returns:
-      A tuple of results.
-    """
-    return tablet_obj.mquery(
-        'vt_test_keyspace',
-        'select id, msg, keyspace_id from %s where id=%d' %
-        (table, mid))
-
-  def _check_value(self, tablet_obj, table, mid, msg, keyspace_id,
-                   should_be_here=True):
-    result = self._get_value(tablet_obj, table, mid)
-    if keyspace_id_type == keyrange_constants.KIT_BYTES:
-      fmt = '%s'
-      keyspace_id = pack_keyspace_id(keyspace_id)
-    else:
-      fmt = '%x'
-    if should_be_here:
-      self.assertEqual(result, ((mid, msg, keyspace_id),),
-                       ('Bad row in tablet %s for id=%d, keyspace_id=' +
-                        fmt + ', row=%s') % (tablet_obj.tablet_alias, mid,
-                                             keyspace_id, str(result)))
-    else:
-      self.assertEqual(
-          len(result), 0,
-          ('Extra row in tablet %s for id=%d, keyspace_id=' +
-           fmt + ': %s') % (tablet_obj.tablet_alias, mid, keyspace_id,
-                            str(result)))
-
-  # _is_value_present_and_correct tries to read a value.
-  # if it is there, it will check it is correct and return True if it is.
-  # if not correct, it will self.fail.
-  # if not there, it will return False.
-  def _is_value_present_and_correct(
-      self, tablet_obj, table, mid, msg, keyspace_id):
-    result = self._get_value(tablet_obj, table, mid)
-    if not result:
-      return False
-    if keyspace_id_type == keyrange_constants.KIT_BYTES:
-      fmt = '%s'
-      keyspace_id = pack_keyspace_id(keyspace_id)
-    else:
-      fmt = '%x'
-    self.assertEqual(result, ((mid, msg, keyspace_id),),
-                     ('Bad row in tablet %s for id=%d, '
-                      'keyspace_id=' + fmt) % (
-                          tablet_obj.tablet_alias, mid, keyspace_id))
-    return True
 
   def _insert_startup_values(self):
     # row covered by shard -40 (should be merged).
@@ -247,7 +173,7 @@ index by_msg (msg)
   def test_merge_sharding(self):
     utils.run_vtctl(['CreateKeyspace',
                      '--sharding_column_name', 'keyspace_id',
-                     '--sharding_column_type', keyspace_id_type,
+                     '--sharding_column_type', base_sharding.keyspace_id_type,
                      'test_keyspace'])
 
     shard_0_master.init_tablet('master', 'test_keyspace', '-40')
@@ -323,7 +249,7 @@ index by_msg (msg)
         'Partitions(master): -40 40-80 80-\n'
         'Partitions(rdonly): -40 40-80 80-\n'
         'Partitions(replica): -40 40-80 80-\n',
-        keyspace_id_type=keyspace_id_type,
+        keyspace_id_type=base_sharding.keyspace_id_type,
         sharding_column_name='keyspace_id')
 
     # copy the schema
@@ -426,7 +352,7 @@ index by_msg (msg)
                              'Partitions(master): -40 40-80 80-\n'
                              'Partitions(rdonly): -80 80-\n'
                              'Partitions(replica): -40 40-80 80-\n',
-                             keyspace_id_type=keyspace_id_type,
+                             keyspace_id_type=base_sharding.keyspace_id_type,
                              sharding_column_name='keyspace_id')
 
     # now serve replica from the split shards
@@ -436,7 +362,7 @@ index by_msg (msg)
                              'Partitions(master): -40 40-80 80-\n'
                              'Partitions(rdonly): -80 80-\n'
                              'Partitions(replica): -80 80-\n',
-                             keyspace_id_type=keyspace_id_type,
+                             keyspace_id_type=base_sharding.keyspace_id_type,
                              sharding_column_name='keyspace_id')
 
     # now serve master from the split shards
@@ -446,7 +372,7 @@ index by_msg (msg)
                              'Partitions(master): -80 80-\n'
                              'Partitions(rdonly): -80 80-\n'
                              'Partitions(replica): -80 80-\n',
-                             keyspace_id_type=keyspace_id_type,
+                             keyspace_id_type=base_sharding.keyspace_id_type,
                              sharding_column_name='keyspace_id')
     utils.check_tablet_query_service(self, shard_0_master, False, True)
     utils.check_tablet_query_service(self, shard_1_master, False, True)
