+++ test/merge_sharding_bytes.py
@@ -8,10 +8,11 @@
 
 from vtdb import keyrange_constants
 
+import base_sharding
 import merge_sharding
 import utils
 
 
 if __name__ == '__main__':
-  merge_sharding.keyspace_id_type = keyrange_constants.KIT_BYTES
+  base_sharding.keyspace_id_type = keyrange_constants.KIT_BYTES
   utils.main(merge_sharding)
